+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: mnascime <mnascime@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/28 13:13:08 by mnascime          #+#    #+#             */
-/*   Updated: 2023/11/06 14:20:17 by paugonca         ###   ########.fr       */
+/*   Updated: 2023/11/13 18:21:20 by paugonca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ int		ft_printf(const char *s, ...);
 int		ft_putchar(char c);
 int		ft_putnbr_base(int str, char vari);
 int		ft_putnbr_p(unsigned long long nb);
-size_t	ft_strlen(const char *s);
+int		ft_strlen(const char *s);
 char	**ft_split(char const *s, char c);
 char	**ft_split_spaces(char const *s);
 void	*ft_memset(void *s, int c, size_t n);
